+++ bin/handlers.mli
@@ -5,6 +5,7 @@ val freight_inspect : State.t -> unit
 val freight_help : State.t -> unit
 val freight_run : State.t -> unit
 val freight_run_all : State.t -> unit
+val freight_jump_run_all : State.t -> int -> unit
 val freight_view_run_all : State.t -> int -> unit
 val freight_view : State.t -> string -> unit
 val freight_history : State.t -> unit